fixed details and reset form
