+++ js/server.js
@@ -1,6 +1,9 @@
 const DATA = 'https://26.javascript.pages.academy/keksobooking/data';
 const SERVER = 'https://26.javascript.pages.academy/keksobooking';
 
+const INTERNAL_SERVER_ERROR = 500;
+const VERSION_NOT_SUPPORTED_ERROR = 505;
+
 const createLoader = (onSuccess, onFail) => fetch(
   DATA,
   {
@@ -31,7 +34,7 @@ const sendForm = (onSuccess, onFail, body) => fetch(
   .then((response) => {
     if (response.ok) {
       onSuccess('Ваше объявление успешно размещено!');
-    } else if (response.status >= 500 && response.status <= 505) {
+    } else if (response.status >= INTERNAL_SERVER_ERROR && response.status <= VERSION_NOT_SUPPORTED_ERROR) {
       onFail('Не удалось получить данные с сервера. Попробуйте ещё раз!');
     } else {
       onFail('Не удалось отправить форму. Попробуйте ещё раз!');
