+++ js/slider.js
@@ -24,11 +24,9 @@ noUiSlider.create(sliderElement, {
 
 type.addEventListener('change', () => {
   sliderElement.noUiSlider.set(minTypePrice[type.value]);
-  valueElement.value = sliderElement.noUiSlider.get();
-  // eslint-disable-next-line no-console
-  console.log(minTypePrice[type.value]);
+  valueElement.value = Math.floor(sliderElement.noUiSlider.get());
 });
 
 sliderElement.noUiSlider.on('update', () => {
-  valueElement.value = sliderElement.noUiSlider.get();
+  valueElement.value = Math.floor(sliderElement.noUiSlider.get());
 });
